removed unnecessary type conversion

+++ rpm.go
@@ -248,7 +248,7 @@ func (r *RPM) writeGenIndexes(h *index) {
 	h.Add(tagSize, entry([]int32{int32(r.payloadSize)}))
 	h.Add(tagName, entry(r.Name))
 	h.Add(tagVersion, entry(r.Version))
-	h.Add(tagEpoch, entry([]uint32{uint32(r.Epoch)}))
+	h.Add(tagEpoch, entry([]uint32{r.Epoch}))
 	h.Add(tagSummary, entry(r.Summary))
 	h.Add(tagDescription, entry(r.Description))
 	h.Add(tagBuildHost, entry(r.BuildHost))
